+++ RevitAddin.DA.Tester/Services/DesignAutomationController.cs
@@ -40,6 +40,9 @@ namespace RevitAddin.DA.Tester.Services
 
             Console.WriteLine($"UI:\t{UI.IsValid()}");
             Console.WriteLine("----------------------------------------");
+            Console.WriteLine($"Shape:\t{typeof(ricaun.Revit.DB.Shape.Colors).Assembly}");
+            Console.WriteLine($"Shape Location:\t{typeof(ricaun.Revit.DB.Shape.Colors).Assembly.Location}");
+            Console.WriteLine("----------------------------------------");
 
             return true;
         }