+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Features
 - Add `DesignAutomation` and `IDesignAutomation` interface.
 - Add `DesignAutomationLoadVersion` to load the right version of the addin.
+- Add `AssemblyResolve` in the `DesignAutomationLoadVersion` to load dependencies in the right version.
 
 ## [1.3.1] / 2025-06-15 - 2025-08-27
 - Add `AddInId` in the output. #9